update: 添加charts
1.添加pie charts
2.添加bar charts

+++ app/src/main/java/com/example/mymoviememoir/utils/MyValueFormmater.java
@@ -5,6 +5,6 @@ import com.github.mikephil.charting.formatter.ValueFormatter;
 public class MyValueFormmater extends ValueFormatter {
     @Override
     public String getFormattedValue(float value) {
-        return String.valueOf(value * 100) + "%";
+        return String.valueOf((int)(value * 100)) + "%";
     }
 }
